packages/agent-failsafe/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-failsafe"
version = "0.1.0"
description = "FailSafe governance bridge for Agent Governance Toolkit"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["ai", "agents", "governance", "failsafe"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pyyaml>=6.0",
    "agent-os-kernel",
    "agent-sre",
    "agentmesh-platform",
    "agent-hypervisor",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.8"]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_failsafe"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "TCH"]
ignore = ["E501", "E741"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]
"tests/*" = ["F401", "E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]
